requirements.txt
netifaces==0.10.4
progressbar2==3.5.0
zeroconf==0.17.4